+++ apps/app/vite.config.ts
@@ -13,6 +13,5 @@ export default createViteConfig({
       devMode: true,
       pluginNodeModulesDir: '../../node_modules'
     })
-  ],
-  devtools: true
+  ]
 });