+++ docs/.vitepress/components/redirect.vue
@@ -3,8 +3,9 @@
 </template>
 <script lang="ts" setup>
   if (typeof window !== 'undefined') {
-    const pathname = window.location.pathname;
-    const path = pathname.replace('redirect.html', 'lcdp/designer/');
-    window.location.replace(path);
+    // const pathname = window.location.pathname;
+    // const path = pathname.replace('redirect.html', 'lcdp/designer/');
+    // window.location.replace(path);
+    window.location.replace('https://lcdp.vtj.pro/');
   }
 </script>