+++ packages/designer/src/components/widgets/index.ts
@@ -26,6 +26,9 @@ import Raw from './raw/index.vue';
 import Docs from './docs/index.vue';
 import About from './about/index.vue';
 import ProjectConfig from './project-config/index.vue';
+import Market from './market/index.vue';
+import Templates from './templates/index.vue';
+import UserAvatar from './user-avatar/index.vue';
 
 export const widgets = {
   Logo,
@@ -55,5 +58,8 @@ export const widgets = {
   Raw,
   Docs,
   About,
-  ProjectConfig
+  ProjectConfig,
+  Market,
+  Templates,
+  UserAvatar
 };