+++ packages/designer/src/framework/simulator.ts
@@ -18,6 +18,7 @@ import {
   mockApis,
   mockCleanup
 } from '@vtj/renderer';
+import html2canvas from 'html2canvas';
 import { logger } from '@vtj/utils';
 import { Renderer } from './renderer';
 import { Designer } from './designer';
@@ -128,14 +129,27 @@ export class Simulator extends Base {
        <meta name="viewport"
              content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0,viewport-fit=cover"/>
          <style>
-            html, body, #app {
-              padding: 0;
-              margin: 0;
-              min-height: 100vh;
-              width: 100%;
-              height: 100%;
-              background: #fff;
-            }
+          html,
+          body,
+          #app {
+            margin: 0;
+            min-height: 100vh;
+            width: 100%;
+            height: 100%;
+            font-size: 14px;
+          }
+          #app {
+            background: var(--el-bg-color, #fff);
+            box-sizing: border-box;
+          }
+          #app.is-page {
+             padding: 10px;
+          }
+            #app.is-page.is-pure {
+             background-color: var(--el-fill-color-light, #f5f7fa);
+             padding: 0;
+          }
+
          </style>
        ${createAssetsCss(css)}
        </head>
@@ -164,7 +178,7 @@ export class Simulator extends Base {
     this.renderer?.dispose();
     this.renderer = null;
     const cw = this.contentWindow as any;
-    const { assets, service, current, provider } = this.engine;
+    const { assets, service, current, provider, project } = this.engine;
     const materialMap = provider.materials || {};
     const materials: Material[] = [];
     for (const name of materialExports) {
@@ -183,7 +197,7 @@ export class Simulator extends Base {
     this.devtools.init(cw, this.engine);
     this.renderer = new Renderer(env, service, provider, this.designer.value);
     if (current.value) {
-      this.renderer.render(current.value);
+      this.renderer.render(current.value, project.value?.currentFile);
       this.rendered.value = Symbol();
     }
     this.triggerReady();
@@ -269,9 +283,23 @@ export class Simulator extends Base {
     this.renderer?.dispose();
     const current = this.engine.current.value;
     if (current) {
-      this.renderer?.render(current);
+      this.renderer?.render(current, this.engine.project.value?.currentFile);
       this.rendered.value = Symbol();
     }
+  }
+
+  capture() {
+    return new Promise((resolve, reject) => {
+      if (!this.contentWindow) return reject(null);
+      const body = this.contentWindow.document.body;
+      html2canvas(body, {
+        allowTaint: true
+      })
+        .then((canvas) => {
+          resolve(canvas);
+        })
+        .catch(reject);
+    });
   }
 
   dispose() {