+++ packages/renderer/src/components/index.ts
@@ -1,0 +1,3 @@
+import Startup from './Startup.vue';
+
+export { Startup };