+++ packages/renderer/src/index.ts
@@ -5,3 +5,5 @@ export * from './provider';
 export * from './render';
 export * from './services';
 export * from './hooks';
+export * from './plugins';
+export * from './components';